Scaffolding for remaining interceptor arguments.
