+++ Src/SDK/WorkflowClient/Temporal.WorkflowClient.Interceptors/public/RequestCancellation.cs
@@ -1,0 +1,24 @@
+﻿namespace Temporal.WorkflowClient.Interceptors
+{
+    public static class RequestCancellation
+    {
+        public record Arguments();
+
+        public class Result : IWorkflowChainBindingResult
+        {
+            public Result(string workflowChainId)
+            {
+                ValidateWorkflowProperty.ChainId.Bound(workflowChainId);
+                WorkflowChainId = workflowChainId;
+            }
+
+            public string WorkflowChainId { get; }
+
+            public bool TryGetBoundWorkflowChainId(out string workflowChainId)
+            {
+                workflowChainId = WorkflowChainId;
+                return (workflowChainId != null);
+            }
+        }
+    }
+}
